app/requirements.locate.txt
pubnub>=4.0.13
python-dateutil
scipy